feat: Restructure code
