+++ src/helpers/relatedProducts.ts
